Magic3 and HealthGauge
The player now has access to the final type of magic, the light version of which restores health while the dark version drains health for a temporary damage buff. The gray version both heals the player and provides a damage buff. The player now also has a health bar that is updated as their health changes.